# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holy Hymns API",
    "version": "1.0.0",
    "description": "Self-hosted lyrics catalogue. Session authorization is checked on every protected request. All timestamps UTC. Public responses never contain drafts."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Local development"
    }
  ],
  "paths": {
    "/songs": {
      "get": {
        "summary": "Published songs, ranked bilingual search",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "letter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "title",
                "recent"
              ],
              "default": "title"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 30,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "maximum": 100000
            }
          }
        ]
      }
    },
    "/songs/{id}": {
      "get": {
        "summary": "Read a published song",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/categories": {
      "get": {
        "summary": "List browsing categories",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/config": {
      "get": {
        "summary": "Read public app configuration and revision",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "summary": "Receive public content revision invalidations",
        "security": [],
        "responses": {
          "200": {
            "description": "SSE event: content; id: durable revision; data: {revision:number}. Reconcile on reconnect.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/favorites": {
      "get": {
        "summary": "List own available favorite song identifiers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/me/favorites/{id}": {
      "put": {
        "summary": "Save favorite",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "delete": {
        "summary": "Remove favorite",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/admin/songs": {
      "get": {
        "summary": "List draft/editor snapshots; admin or owner",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 30,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "maximum": 100000
            }
          }
        ],
        "x-required-role": "admin"
      },
      "post": {
        "summary": "Create unpublished song; admin or owner",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SongInput"
              }
            }
          }
        },
        "x-required-role": "admin"
      }
    },
    "/admin/songs/{id}": {
      "get": {
        "summary": "Read editor snapshot; admin or owner",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-required-role": "admin"
      },
      "put": {
        "summary": "Save draft with version guard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SongUpdateInput"
              }
            }
          }
        },
        "x-required-role": "admin"
      }
    },
    "/admin/songs/{id}/publish": {
      "post": {
        "summary": "Publish atomically and advance public revision",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "publish"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Version"
              }
            }
          }
        },
        "x-required-role": "admin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable account session, not a provider identity token."
      }
    },
    "schemas": {
      "Song": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "titleMalayalam": {
            "type": "string"
          },
          "lyricsMalayalam": {
            "type": "string"
          },
          "lyricsManglish": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "credits": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaLink"
            }
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "featured": {
            "type": "boolean"
          },
          "version": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sourceUrl": {
            "type": "string"
          },
          "reviewNotes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Administrative review notes; omitted from public song responses."
          }
        },
        "required": [
          "id",
          "title",
          "titleMalayalam",
          "lyricsMalayalam",
          "lyricsManglish",
          "aliases",
          "credits",
          "links",
          "categoryIds",
          "featured",
          "version",
          "status",
          "updatedAt"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SongList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
...
    }
  }
}